[FIX] Fuse Clock and Date to a single component
Display Clock, Week & Date in a better way. Combined Clock.tsx and DateDisplay.tsx. Cleaned up in Infoscreen.tsx
