--- frontend/info-skjerm-2026/src/components/DateDisplay.tsx
@@ -1,30 +1,0 @@
-import React from "react";
-
-function getWeekNumber(date: Date): number {
-  const d = new Date(
-    Date.UTC(date.getFullYear(), date.getMonth(), date.getDate()),
-  );
-  const day = d.getUTCDay() || 7;
-  d.setUTCDate(d.getUTCDate() + 4 - day);
-  const yearStart = new Date(Date.UTC(d.getUTCFullYear(), 0, 1));
-  return Math.ceil(((d.getTime() - yearStart.getTime()) / 86400000 + 1) / 7);
-}
-
-const DateDisplay: React.FC = () => {
-  const now = new Date();
-  const dd = String(now.getDate()).padStart(2, "0");
-  const mm = String(now.getMonth() + 1).padStart(2, "0");
-  const yy = String(now.getFullYear()).slice(-2);
-  const week = getWeekNumber(now);
-
-  return (
-    <div>
-      <h1 className="text-center text-5xl">
-        {dd}.{mm}.{yy}
-      </h1>
-      <h1 className="text-center text-5xl">Uke {week}</h1>
-    </div>
-  );
-};
-
-export default DateDisplay;
